+++ Illuminati/Extensions/UIView+TopBorder.swift
@@ -10,7 +10,7 @@ import UIKit
 import QuartzCore
 
 extension UIView {
-    func addTopBorder(width: Float, height: Float, color: UIColor) -> Self {
+    func addTopBorder(width: CGFloat, height: CGFloat, color: UIColor) -> Self {
         let topBorder = CALayer()
         topBorder.frame = CGRect(x: 0, y: 0, width: width, height: height)
         topBorder.backgroundColor = color.CGColor
