+++ Illuminati/SecretViewController.swift
@@ -65,7 +65,7 @@ class SecretViewController: UIViewController, UITableViewDataSource, UITableView
         frame.size.height = min(max(100, newHeight), maxHeight)
         
         if (frame.size.height != dv.secretView.frame.height) {
-            dv.secretView.compactness = frame.size.height / maxHeight
+            dv.secretView.compactness = Float(frame.size.height / maxHeight)
             dv.secretView.frame = frame
         }
     }
